% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{align}
\alias{align}
\alias{align_top}
\alias{align_bottom}
\alias{align_left}
\alias{align_right}
\title{Align the top / bottom / left / right edge of \code{a} to \code{b}}
\usage{
align_top(scene, a, b)

align_bottom(scene, a, b)

align_left(scene, a, b)

align_right(scene, a, b)
}
\arguments{
\item{scene}{An \code{inc_scene}.}

\item{a, b}{Plot indices or element ids.}
}
\description{
Align the top / bottom / left / right edge of \code{a} to \code{b}
}
