% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialize.R
\name{incantation-io}
\alias{incantation-io}
\alias{write_incantation}
\alias{apply_incantation}
\title{Write / read an edit manifest as JSON}
\usage{
write_incantation(scene, path)

apply_incantation(scene, path)
}
\arguments{
\item{scene}{An \code{inc_scene}. @param path File path of a manifest.}

\item{path}{File path.}
}
\description{
Write / read an edit manifest as JSON

Re-apply a manifest's operations to a scene
}
