% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure.R
\name{is_empty_element}
\alias{is_empty_element}
\title{Is an element empty (a spacer or blank grob)?}
\usage{
is_empty_element(scene, id)
}
\arguments{
\item{scene}{An \code{inc_scene}.}

\item{id}{Element id.}
}
\description{
Is an element empty (a spacer or blank grob)?
}
