% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure.R
\name{plot_label}
\alias{plot_label}
\title{The y-axis title text of a plot, used to identify which plot an index points to}
\usage{
plot_label(scene, ref)
}
\arguments{
\item{scene}{An \code{inc_scene}.}

\item{ref}{Plot index or element id.}
}
\description{
The y-axis title text of a plot, used to identify which plot an index points to
}
