% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{render}
\alias{render}
\title{Render the edited figure (optionally with debug boxes) as a drawable gtable}
\usage{
render(scene, debug = FALSE)
}
\arguments{
\item{scene}{An \code{inc_scene}.}

\item{debug}{Draw element bounding boxes and ids for inspection?}
}
\description{
Render the edited figure (optionally with debug boxes) as a drawable gtable
}
