% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select_id}
\alias{select_id}
\title{Select an element by its exact id}
\usage{
select_id(scene, id)
}
\arguments{
\item{scene}{An \code{inc_scene}.}

\item{id}{Element id, e.g. "plot-3/panel" or "inset-2".}
}
\description{
Select an element by its exact id
}
