% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select_panel}
\alias{select_panel}
\title{Select a subplot's panel (its data region)}
\usage{
select_panel(scene, index)
}
\arguments{
\item{scene}{An \code{inc_scene}.}

\item{index}{Integer plot index as the user composed it.}
}
\description{
Select a subplot's panel (its data region)
}
