% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select_plot}
\alias{select_plot}
\title{Select a whole subplot by its plot index}
\usage{
select_plot(scene, index)
}
\arguments{
\item{scene}{An \code{inc_scene}.}

\item{index}{Integer plot index as the user composed it.}
}
\description{
Select a whole subplot by its plot index
}
