% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svg-transform.R
\name{svg_rotate}
\alias{svg_rotate}
\title{Rotate selected SVG primitives around their collective centre}
\usage{
svg_rotate(x, angle)
}
\arguments{
\item{x}{An \code{inc_svg} with an active selection.}

\item{angle}{Clockwise angle in degrees.}
}
\description{
Rotate selected SVG primitives around their collective centre
}
