% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svg-transform.R
\name{svg_scale}
\alias{svg_scale}
\title{Scale selected SVG primitives around their collective centre}
\usage{
svg_scale(x, sx = 1, sy = sx)
}
\arguments{
\item{x}{An \code{inc_svg} with an active selection.}

\item{sx, sy}{Horizontal and vertical scale factors.}
}
\description{
Scale selected SVG primitives around their collective centre
}
