% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svg-transform.R
\name{svg_zorder}
\alias{svg_zorder}
\title{Move selected SVG primitives to the front or back of their parent}
\usage{
svg_zorder(x, where = c("front", "back"))
}
\arguments{
\item{x}{An \code{inc_svg} with an active selection.}

\item{where}{"front" or "back".}
}
\description{
Move selected SVG primitives to the front or back of their parent
}
