% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svg-render.R
\name{write_svg}
\alias{write_svg}
\title{Write an edited SVG snapshot}
\usage{
write_svg(x, path)
}
\arguments{
\item{x}{An \code{inc_svg}.}

\item{path}{Output SVG path.}
}
\description{
Write an edited SVG snapshot
}
